- Size string initializers to fit their terminator, drop unused variables, and   convert remaining internal K&R definitions in extfuncs - Default the aspect ratio to 1.0 when no active graphics device reports one

+++ src/gdm/getdevicerat.c
@@ -31,7 +31,7 @@ getdeviceratio(float *ratio) {
             *ratio = fmin(*ratio, ratio2);
         }
     }
-    if (*ratio == VLARGE) {
+    if (*ratio == (float) VLARGE) {
         *ratio = 1.0;
     }
